add AuthContext to provide user infor for entire app, and use conditional rendering to control AuthStack or AppStack gets shown, so that PetsContext only needs to wrap AppStack
