+++ PetDiaryApp/AuthStack.js
@@ -1,0 +1,28 @@
+import React, { useEffect, useState } from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+import LoginScreen from "./screens/LoginScreen";
+import SignupScreen from "./screens/SignupScreen";
+import { colors } from "./styles";
+
+const Stack = createNativeStackNavigator();
+
+const AuthStack = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        screenOptions={{
+          headerStyle: { backgroundColor: colors.backgroundColor },
+          headerTintColor: colors.defaultTextColor,
+        }}
+        initialRouteName="Signup"
+      >
+        <Stack.Screen name="Login" component={LoginScreen} />
+        <Stack.Screen name="Signup" component={SignupScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
+
+export default AuthStack;
